+++ src/backend/main/go.main/facilities/REST/RESTfacility.go
@@ -21,6 +21,7 @@ type (
 		LocalsIdentities(user_id string) (identities []LocalIdentity, err error)
 		SuggestRecipients(user_id, query_string string) (suggests []RecipientSuggestion, err error)
 		ContactIdentities(user_id, contact_id string) (identities []ContactIdentity, err error)
+		GetSettings(user_id string) (settings *Settings, err error)
 		//messages
 		GetMessagesList(filter IndexSearch) (messages []*Message, totalFound int64, err error)
 		GetMessage(user_id, message_id string) (message *Message, err error)